+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Most recent version is listed first.
 
 ## v0.0.40
 - Detect leaks in tests: https://github.com/komuw/ong/pull/232
+- Bugfix; loadshed records latency in milliseconds: https://github.com/komuw/ong/pull/233
 
 ## v0.0.39
 - Remove pid from logs: https://github.com/komuw/ong/pull/230